requirements-dev.txt
flexmock==0.9.4
pep8==1.6.2
PyHamcrest==1.8.5
